+++ framework/Cargo.toml
@@ -80,7 +80,6 @@ ibc-host = { package = "abstract-ibc-host", path = "contracts/native/ibc-host" }
 
 abstract-account = { path = "contracts/account", default-features = false }
 
-abstract-ica = { path = "packages/abstract-ica" }
 abstract-sdk = { version = "0.23.0", path = "packages/abstract-sdk" }
 abstract-testing = { version = "0.23.0", path = "packages/abstract-testing" }
 abstract-std = { version = "0.23.0", path = "packages/abstract-std" }
