+++ framework/contracts/native/ica-client/src/chain_types/evm.rs
@@ -1,8 +1,8 @@
-use abstract_ica::EVM_NOTE_ID;
 use abstract_sdk::{
     feature_objects::{AnsHost, RegistryContract},
     Resolve,
 };
+use abstract_std::ica_client::EVM_NOTE_ID;
 use abstract_std::objects::{module::ModuleInfo, ChannelEntry, ContractEntry, TruncatedChainId};
 use cosmwasm_std::{
     wasm_execute, Addr, Binary, Coin, CosmosMsg, Deps, Env, HexBinary, QuerierWrapper, WasmMsg,
@@ -95,8 +95,10 @@ pub fn send_funds(
 }
 
 fn evm_note_addr(vc: &RegistryContract, querier: &QuerierWrapper) -> IcaClientResult<Addr> {
-    let evm_note_entry =
-        ModuleInfo::from_id(EVM_NOTE_ID, abstract_ica::POLYTONE_EVM_VERSION.parse()?)?;
+    let evm_note_entry = ModuleInfo::from_id(
+        EVM_NOTE_ID,
+        abstract_std::ica_client::POLYTONE_EVM_VERSION.parse()?,
+    )?;
 
     vc.query_module(evm_note_entry, querier)?
         .reference
