--- framework/packages/abstract-ica/Cargo.toml
@@ -1,41 +1,0 @@
-[package]
-name = "abstract-ica"
-version = { workspace = true }
-authors = { workspace = true }
-edition = { workspace = true }
-description = "Abstract Interchain Account (ICA) Types"
-license = { workspace = true }
-readme = "README.md"
-repository = "https://github.com/AbstractSDK/abstract"
-
-
-[features]
-# for quicker tests, cargo test --lib
-
-[dependencies]
-cosmwasm-std = { workspace = true, features = ["iterator"] }
-cosmwasm-schema = { workspace = true }
-thiserror = { workspace = true }
-schemars = { workspace = true }
-cw-storage-plus = { workspace = true }
-cw-ownable = { workspace = true }
-serde = { workspace = true }
-map-macro = "0.3.0"
-
-cw-orch = { workspace = true }
-abstract-sdk = { workspace = true }
-
-# EVM
-polytone-evm = { version = "0.1.0" }
-alloy = { version = "0.2.1", default-features = false }
-alloy-sol-types = { version = "0.7.7", default-features = false }
-
-[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-workspace-hack = { version = "0.1", path = "../../workspace-hack" }
-
-[dev-dependencies]
-speculoos = { workspace = true }
-rstest = { workspace = true }
-anyhow = { workspace = true }
-abstract-testing = { path = "../abstract-testing" }
-cw20-base = { workspace = true }
