+++ framework/packages/abstract-std/Cargo.toml
@@ -37,6 +37,9 @@ cw-blob = { workspace = true, default-features = false }
 ## Stringify function names
 function_name = { version = "0.3.0" }
 
+polytone-evm = { version = "0.0.2" }
+map-macro = "0.3.0"
+
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 workspace-hack = { version = "0.1", path = "../../workspace-hack" }
